remove casting in tf

+++ tf_model_profiling.py
@@ -11,7 +11,7 @@ import tensorflow as tf
 class Dataset:
     def __init__(self, steps, batch_size):
         self.data = dict()
-        self.data['state'] = {'pov': random.randint(0, 255, size=(steps*batch_size, 256, 256, 12), dtype='uint8'),
+        self.data['state'] = {'pov': random.randint(0, 255, size=(steps*batch_size, 256, 256, 12), dtype='float32'),
                               'angles': random.uniform(-2*np.pi, -2*np.pi, size=(steps*batch_size, 6))}
         self.data['action'] = np.ones(steps*batch_size, dtype='int32')
         self.data['reward'] = np.ones(steps*batch_size, dtype='float32')
